Add files via upload
DESCRIPTION

Create a servlet-based web application that shows a form to enter product details. Capture the details in a servlet and then display the data that was entered.

Background of the problem statement:
As a part of developing an ecommerce web application, you have to prototype a form for adding products into the system entered by the users. There is no database involved here, so
the product is just captured and displayed without storing it anywhere.

You must use the following:

Eclipse as the IDE
Apache Tomcat as the web server
JSP pages for the front end
Servlets for backend processing
Following requirements should be met:

There will be a Product class to store the details of the product
The servlet will put the form data into the Product class and store it in the session
A second JSP will take the product info from the session and display it to the user
The step-by-step process involved in completing this task should be documented